omega/oaf/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
idna==3.4
packageurl-python==0.10.4
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
